refactor :: splashActivity 네이밍 변경
