+++ app/src/main/java/com/example/reverie2/MainActivity.kt
@@ -1,11 +1,14 @@
-package com.example.reverie_v2
-
+package com.example.reverie2
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.*
-import com.example.reverie_v2.ui.feature.splash.SplashScreen
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import com.example.reverie2.ui.feature.splash.splashScreen
 import kotlinx.coroutines.delay
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,7 +20,7 @@ class MainActivity : ComponentActivity() {
                 delay(2500)
                 showSplash = false
             }
-            SplashScreen()
+            splashScreen()
         }
     }
 }
